throw exception in user delete

+++ app/Services/UserServices.php
@@ -153,6 +153,10 @@ class UserServices
         try {
             $user = User::find($userId);
             if ($user) {
+                $projects = Project::where('manage_by',$userId)->get();
+                if(isset($projects)){
+                    throw new Exception('This User Manage Few Projects. Soo You can not Delete this User.');
+                }
                 // unlink('avatars/' . $user->avtar);
                 $user->delete();
                 $response = ['status' => 'Success', 'message' => 'User Delete Successfully.'];
